Move test_power_balancer to its own file
- Some minor changes to support --skip-launch.
