+++ integration/test/geopm_test_integration.py
@@ -747,108 +747,6 @@ class TestIntegration(unittest.TestCase):
             # TODO Checks on the maximum power computed during the run?
             # TODO Checks to see how much power was left on the table?
 
-    @util.skip_unless_run_long_tests()
-    @util.skip_unless_batch()
-    def test_power_balancer(self):
-        name = 'test_power_balancer'
-        num_node = 4
-        num_rank = 16
-        loop_count = 500
-        # Require that the balancer moves the maximum dgemm runtime at
-        # least 1/4 the distance to the mean dgemm runtime under the
-        # governor.
-        margin_factor =  0.25
-        app_conf = geopmpy.io.BenchConf(name + '_app.config')
-        self._tmp_files.append(app_conf.get_path())
-        app_conf.append_region('dgemm-imbalance', 8.0)
-        app_conf.append_region('all2all', 0.05)
-        app_conf.set_loop_count(loop_count)
-
-        # Update app config with imbalance
-        alloc_nodes = geopm_test_launcher.TestLauncher.get_alloc_nodes()
-        for nn in range(len(alloc_nodes) // 2):
-            app_conf.append_imbalance(alloc_nodes[nn], 0.5)
-
-        fam, mod = geopm_test_launcher.get_platform()
-        if fam == 6 and mod == 87:
-            # budget for KNL
-            power_budget = 130
-        else:
-            power_budget = 200
-        self._options = {'power_budget': power_budget}
-        gov_agent_conf_path = name + '_gov_agent.config'
-        bal_agent_conf_path = name + '_bal_agent.config'
-        self._tmp_files.append(gov_agent_conf_path)
-        self._tmp_files.append(bal_agent_conf_path)
-
-        agent_list = ['power_governor', 'power_balancer']
-        path_dict = {'power_governor': gov_agent_conf_path, 'power_balancer': bal_agent_conf_path}
-        agent_runtime = dict()
-        for agent in agent_list:
-            agent_conf = geopmpy.io.AgentConf(path_dict[agent], agent, self._options)
-            run_name = '{}_{}'.format(name, agent)
-            report_path = '{}.report'.format(run_name)
-            trace_path = '{}.trace'.format(run_name)
-            launcher = geopm_test_launcher.TestLauncher(app_conf, agent_conf, report_path,
-                                                        trace_path, time_limit=2700)
-            launcher.set_num_node(num_node)
-            launcher.set_num_rank(num_rank)
-            launcher.write_log(run_name, 'Power cap = {}W'.format(power_budget))
-            launcher.run(run_name)
-
-            self._output = geopmpy.io.AppOutput(report_path, trace_path + '*')
-            node_names = self._output.get_node_names()
-            self.assertEqual(num_node, len(node_names))
-
-            power_limits = []
-            # Total power consumed will be Socket(s) + DRAM
-            for nn in node_names:
-                tt = self._output.get_trace_data(node_name=nn)
-
-                first_epoch_index = tt.loc[tt['EPOCH_COUNT'] == 0][:1].index[0]
-                epoch_dropped_data = tt[first_epoch_index:]  # Drop all startup data
-
-                power_data = epoch_dropped_data.filter(regex='ENERGY')
-                power_data['TIME'] = epoch_dropped_data['TIME']
-                power_data = power_data.diff().dropna()
-                power_data.rename(columns={'TIME': 'ELAPSED_TIME'}, inplace=True)
-                power_data = power_data.loc[(power_data != 0).all(axis=1)]  # Will drop any row that is all 0's
-
-                pkg_energy_cols = [s for s in power_data.keys() if 'ENERGY_PACKAGE' in s]
-                dram_energy_cols = [s for s in power_data.keys() if 'ENERGY_DRAM' in s]
-                power_data['SOCKET_POWER'] = power_data[pkg_energy_cols].sum(axis=1) / power_data['ELAPSED_TIME']
-                power_data['DRAM_POWER'] = power_data[dram_energy_cols].sum(axis=1) / power_data['ELAPSED_TIME']
-                power_data['COMBINED_POWER'] = power_data['SOCKET_POWER'] + power_data['DRAM_POWER']
-
-                pandas.set_option('display.width', 100)
-                launcher.write_log(name, 'Power stats from {} {} :\n{}'.format(agent, nn, power_data.describe()))
-
-                # Get final power limit set on the node
-                if agent == 'power_balancer':
-                    power_limits.append(epoch_dropped_data['POWER_LIMIT'][-1])
-
-            if agent == 'power_balancer':
-                avg_power_limit = sum(power_limits) / len(power_limits)
-                self.assertTrue(avg_power_limit <= power_budget)
-
-            max_runtime = float('nan')
-            node_names = self._output.get_node_names()
-            runtime_list = []
-            for node_name in node_names:
-                epoch_data = self._output.get_report_data(node_name=node_name, region='dgemm')
-                runtime_list.append(epoch_data['runtime'].item())
-            if agent == 'power_governor':
-                mean_runtime = sum(runtime_list) / len(runtime_list)
-                max_runtime = max(runtime_list)
-                margin = margin_factor * (max_runtime - mean_runtime)
-
-            agent_runtime[agent] = max(runtime_list)
-
-        self.assertGreater(agent_runtime['power_governor'] - margin,
-                           agent_runtime['power_balancer'],
-                           "governor runtime: {}, balancer runtime: {}, margin: {}".format(
-                               agent_runtime['power_governor'], agent_runtime['power_balancer'], margin))
-
     def test_progress_exit(self):
         """
         Check that when we always see progress exit before the next entry.
